requirements.txt
streamlit
pandas
numpy<2.0
joblib
shap
scikit-learn
xgboost
matplotlib
seaborn
statsmodels
scipy
streamlit-shap
patsy
ucimlrepo
lime

langchain
langchain-community
langchain-core
langchain-google-genai
langchain-experimental
langchain-huggingface
pypdf
python-dotenv
faiss-cpu
datasets
huggingface_hub
langchain-openai
sentence-transformers
open_clip_torch
Pillow
duckduckgo-search
tiktoken
pydantic
